+++ client/app/components/visualizations/EditVisualizationDialog.less
