+++ client/app/components/visualizations/VisualizationName.jsx
