+++ client/app/components/visualizations/VisualizationName.less
