+++ client/app/visualizations/components/Renderer.jsx
